Removed endpoint to add/remove log collector
